requirements.txt
bcrypt==3.2.0
Flask==2.0.1
Flask-JWT-Extended==4.2.1
Flask-SQLAlchemy==2.5.1
marshmallow==3.12.1
pre-commit==2.13.0
PyMySQL==1.0.2
pytest==6.2.4
python-dotenv==0.17.1
